Fix README and follow python coding style
